Optimization, Data write + read scripts
I changed the way we are processing noise reduction, so it is only done once we have an entirely new tensor of values.

This aided in creating the data collection function to collect and label lidar point clouds.

After outputing the labeled data to a csv, created another script to read the data